Restructure repository in preparation of adding QC things.
